requirements.txt
ansible==2.0.1.0
ecdsa==0.13
Jinja2==2.8
lxml==3.5.0
MarkupSafe==0.23
paramiko==1.16.0
prettytable==0.7.2
pycrypto==2.6.1
PyYAML==3.11
requests==2.9.1
xmltodict==0.10.1
